requirements.txt
streamlit
pandas
numpy
google-generativeai
pypdf
